+++ src/methods.js
@@ -12,6 +12,12 @@ const methods = [
     isSuccess: true,
   },
   {
+    name: 'accepted',
+    code: '202',
+    message: 'Accepted',
+    isSuccess: true,
+  },
+  {
     name: 'noContent',
     code: '204',
     message: 'No Content',
@@ -48,9 +54,27 @@ const methods = [
     isSuccess: false,
   },
   {
+    name: 'timeout',
+    code: '408',
+    message: 'Request Timeout',
+    isSuccess: false,
+  },
+  {
+    name: 'conflict',
+    code: '409',
+    message: 'Conflict',
+    isSuccess: false,
+  },
+  {
     name: 'unprocess',
     code: '422',
     message: 'Unprocessable Entity',
+    isSuccess: false,
+  },
+  {
+    name: 'tooManyRequests',
+    code: '429',
+    message: 'Too Many Requests',
     isSuccess: false,
   },
   {